renaming KeyService to KeyRecordService
