+++ src/Hyperledger.Aries/Configuration/ServiceCollectionExtensions.cs
@@ -11,6 +11,7 @@ using Hyperledger.Aries.Features.OpenId4Vc.Vci.Services.Oid4VciService;
 using Hyperledger.Aries.Features.OutOfBand;
 using Hyperledger.Aries.Features.PresentProof;
 using Hyperledger.Aries.Features.RevocationNotification;
+using Hyperledger.Aries.Features.SdJwt.Services.KeyRecordService;
 using Hyperledger.Aries.Features.SdJwt.Services.KeyService;
 using Hyperledger.Aries.Features.SdJwt.Services.SdJwtCredentialService;
 using Hyperledger.Aries.Ledger;
@@ -113,7 +114,7 @@ namespace Microsoft.Extensions.DependencyInjection
         internal static IServiceCollection AddOpenIdDefaultServices(this IServiceCollection builder)
         {
             builder.AddSingleton<IHolder, Holder>();
-            builder.AddSingleton<IKeyService, KeyService>();
+            builder.AddSingleton<IKeyRecordService, KeyRecordService>();
             builder.AddSingleton<ISdJwtCredentialService, DefaultSdJwtCredentialService>();
             builder.AddSingleton<IOid4VciService, DefaultOid4VciService>();
